refactor: QrCodeConfig 파일 패키지 이동
