+++ backend/src/main/java/coffeeshout/room/config/QrCodeConfig.java
@@ -1,4 +1,4 @@
-package coffeeshout.global.config;
+package coffeeshout.room.config;
 
 import com.google.zxing.qrcode.QRCodeWriter;
 import org.springframework.context.annotation.Bean;
